+++ opendc-compute/opendc-compute-simulator/src/test/kotlin/org/opendc/compute/simulator/SimHostTest.kt
@@ -30,11 +30,14 @@ import org.junit.jupiter.api.assertAll
 import org.opendc.compute.api.*
 import org.opendc.compute.service.driver.Host
 import org.opendc.compute.service.driver.HostListener
-import org.opendc.simulator.compute.kernel.SimFairShareHypervisorProvider
+import org.opendc.simulator.compute.SimBareMetalMachine
+import org.opendc.simulator.compute.kernel.SimFairShareHypervisor
 import org.opendc.simulator.compute.model.MachineModel
 import org.opendc.simulator.compute.model.MemoryUnit
 import org.opendc.simulator.compute.model.ProcessingNode
 import org.opendc.simulator.compute.model.ProcessingUnit
+import org.opendc.simulator.compute.power.ConstantPowerModel
+import org.opendc.simulator.compute.power.SimplePowerDriver
 import org.opendc.simulator.compute.workload.SimTrace
 import org.opendc.simulator.compute.workload.SimTraceFragment
 import org.opendc.simulator.compute.workload.SimTraceWorkload
@@ -67,16 +70,16 @@ internal class SimHostTest {
     fun testOvercommitted() = runBlockingSimulation {
         val duration = 5 * 60L
         val engine = FlowEngine(coroutineContext, clock)
-        val random = SplittableRandom(1)
+        val machine = SimBareMetalMachine(engine, machineModel, SimplePowerDriver(ConstantPowerModel(0.0)))
+        val hypervisor = SimFairShareHypervisor(engine, null, SplittableRandom(1))
         val host = SimHost(
             uid = UUID.randomUUID(),
             name = "test",
-            model = machineModel,
             meta = emptyMap(),
             coroutineContext,
-            engine,
-            SimFairShareHypervisorProvider(),
-            random,
+            clock,
+            machine,
+            hypervisor
         )
         val vmImageA = MockImage(
             UUID.randomUUID(),
@@ -151,16 +154,16 @@ internal class SimHostTest {
     fun testFailure() = runBlockingSimulation {
         val duration = 5 * 60L
         val engine = FlowEngine(coroutineContext, clock)
-        val random = SplittableRandom(1)
+        val machine = SimBareMetalMachine(engine, machineModel, SimplePowerDriver(ConstantPowerModel(0.0)))
+        val hypervisor = SimFairShareHypervisor(engine, null, SplittableRandom(1))
         val host = SimHost(
             uid = UUID.randomUUID(),
             name = "test",
-            model = machineModel,
             meta = emptyMap(),
             coroutineContext,
-            engine,
-            SimFairShareHypervisorProvider(),
-            random
+            clock,
+            machine,
+            hypervisor
         )
         val image = MockImage(
             UUID.randomUUID(),
